+++ nakadion/CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [X]
+
+### CHANGED
+
+- renamed `BatchLine` to `EventStreamBatch` to match Nakadi names
+- renamed `BatchLineStream` to `EventStream` to match Nakadi names
+
 ## [0.23.3] - 2020-05-11
 
 - added metrics for connector